Fix output of author for Google+
When the front page is set to static and a page or post, the author is displayed even when it is set to not display.

The first check should be if it's the home page. Both is_home and is_front_page need to be used. is_home doesn't return true when the front page is static and set to a post.

The next step is to check if the option plus-author is set and it's doesn't have the value of -1, which is the value given when you select not to show an author on the home page.

Signed-off-by: Peter van der Does <github@avirtualhome.com>