Fixing dismissible notice on Multisite
update_site_option works as update_network_option on a multinetwork. i.e. stores data in site_meta not wp_options or wp_*_options. This paired with the fact that checks if the notice was dismissed was at network wide and not site wide level meant that the option was never dismissed