Caching empty sitemaps
By wrapping cached sitemaps we can provide meta information about the status of the sitemap.
This way we can also cache empty or bad sitemaps so we don't have to check logic or do database requests if an invalid sitemap is requested.
Made sure that the index sitemap also flags bad if no sitemaps are found to list.