Enable advanced settings pages when the user visits advanced page
The issue:
When the advanced settings are disabled, but an advanced page is
visited in the browser, the user gets an 'insufficient rights'
error.

This PR makes sure that when the user visits an advanced page,
the advanced settings are enabled, as it's obvious the user needs
to be able to visit the advanced settings.

Fixes #5925