Refactor menu registration
This was needed to provide the ability to register menu items as main menu items for specific capabilities.
Best example is the redirects manager role which will be introduced in WPSEO Premium. This should be accessible for users with the role or capability, but they will not be able to access the normal menu structure.