CS/I18n: fix missing text domain in gettext call
Plugins/themes should use their own text domain. however, in some cases, using the WP Core translation for a string is warranted, like in this case.

The WP Core text-domain is `default` and while all gettext function call have a default value of `default` for the `$domain` parameter, explicitly setting the `$domain` in a gettext function call to `default` documents that the intention for this function call was to use the WP Core translation.

For CS purposes, the line still has to be ignored (for now) as it doesn't use the WPSEO text domain. Once WPSEO upgrades to a more recent version of WordPressCS, the ignore comment can be removed as since WPCS 0.11.0, a plugin can indicate that it uses more than one text-domain.