CS/best practices: add note to replace a number of `parse_url()` calls in the near future
The WP `wp_parse_url()` function did not support the `$component` parameter which `parse_url()` has.

This has been fixed in WP 4.7, so calls to `parse_url()` using the `$component` parameter have not been replaced (yet), but a note has been made with each of them to remind the team to do so in the near future.