CS/cleanup: don't use assignments in conditions - part 2
Assigning the values to be passed to a function call to variables is unnecessary.

If it's done for reasons of documentation, it should be done in separate variable assignments before the function call is made.

If it's done as those variabels are needed/used afterwards, again, it should be done in separate variable assignments before the function call is made.

As neither of these two reasons seems to apply here, I've removed the variable assignment within the function calls.