CS/cleanup: don't use assignments in conditions - part 4-b
This commit deals with a variety of function calls within conditions, where the assignment was a secondary condition.

These assignments have been moved within the first condition and a nested conditional has been added to evaluated the return value of the function call.

For each I've looked at the possible return values from the underlying function and what the code within the condition expects to receive and have chosen the IMHO most appropriate variable test.