CS/cleanup: don't use assignments in conditions/remove conditions - part 5-a
This commit deals with an assignment in condition, where the condition was in actual fact superfluous.

The `get_focus_keyword()` call will always return a string, either with the requested value or term name as the default, so no need to test the return before using it.

Additionally, as the condition would always be true, the code underneath it effectively was unreachable, so might as well be removed.