Always use transient to check if the tables are accessible
Introduce extra functions to better check against transient.
The transient was only used to check if the table was inaccessble at some point.
This commit introduces the transient value when the table is accessble, without an expiry time.

This will prevent any table checks having to be made after a successful one.