Back port WordPress globals to classic editor
This code takes into account 3 scenarios:

1. WordPress without Gutenberg. We then bundle the Gutenberg scripts
using the backporting script.

2. WordPress with Gutenberg, editing in classic editor. We enqueue the
scripts Gutenberg provides and then put the things we need on our
global variable.

3. WordPress with Gutenberg, editing in Gutenberg. We copy the
variables from the Gutenberg globals to our own globals.

As soon as we only support a version of WordPress that includes
Gutenberg we can drop scenario 3.

I use a custom pragma, because that way we don't have to check in the
scripts that use React if React is present where we expect it.