Setup webpack config for the snippet editor
This excludes all packages from babel parsing except yoast-components.
That makes the build faster and also prevents DraftJS from including a
bunch of packages for which we don't have the correct babel plugins
and presets.

I also included JSON files in the webpack config, because we need that
in yoast-components. And it is just nice to be able to load json
files.