Refactor the `WPSEO_Customizer` class
While working on the multi-line function call changes, I noticed that the `WPSEO_Customizer` class contained a lot of duplicate code.

This PR:
* Abstracts the duplicate code to a `private` utility/helper method.
* Moves the default settings used by nearly all of the methods to class properties.
* Adjusts the code in the methods to use the helper methods instead.

While working on this, I noticed one inconsistency: the breadcrumb separator setting has a different ID, `breadcrumbs-sep`, from the associated control `(wpseo-)breadcrumbs-separator`.
In all the other cases, these are in sync.

Please let me know if I should update this PR to let these IDs always be in sync.
In that case, the optional `$id` parameter of the new `add_setting_and_control()` method will be removed.

Also note that the new `add_setting_and_control()` method contains a `$settings_args` parameter which is currently not used as all settings use the default arguments. However, having the parameter there allows for the flexibility to pass different settings arguments if needs be.