CS: remove an unnecessary property declaration
Properties should not be prefixed with an `_` to indicate visibility. This is an outdated PHP4-style habit.

In this case, the property name can not be changed as the property comes from the WP native `WP_List_Table` class which this class extends, but as the property is declared in the parent class and the property declaration in this child class does not overwrite a default value or change the visibility of the property, there is no need to declare it in the child class in the first place, so we may as well remove it.