Pass refreshAnalysis callback instead of app to bindElementEvents
We were passing the app (which is an object) to postDataCollector.bindElementEvents, which was expecting a callback. As a result, further down the road, TinyMCE received an object where it was expecting a function, which led to console errors and making it impossible to paste text. As a solution, we now pass the refreshAnalysis callback instead. To termScraper.bindElementEvents we were still passing the old-skool YoastSEO.app.refresh. This is now replaced by the refreshAnalysis callback as well.