CS: don't use underscore prefixed property names [3]
Properties should not be prefixed with an `_` (to indicate visibility or for whatever other reason). This is an outdated PHP4-style habit.

Renaming the property and instances of it in child classes fixes this.

While the `ORMWrapper` class extends the `ORM` class, the naming of this property does not appear to be dictated by the parent class, so it should be safe to rename it.

Note: This should be considered a BC-break.