[src] QA: consistently use `use` statements
A lot of the time, the new tests did used external classes, without importing them with `use` statements.

As it is good practice to always import all classes used within a namespaced file using `use` statements, this has now been fixed.

This is on the one hand about consistency across the classes, on the other hand about documenting which classes are used in a file.

Only one exception is made and that is for the PHP native `Exception` class as using that in its fully qualified form makes it unambiguous whether the PHP native `Exception` class or a userland `Exception` class has been used.

As for the order of the `use` statements, I've listed those as such:
* Functional classes first in alphabetic order.
* Test related classes after that, again in alphabetic order.
* [If applicable] PHP native classes in alphabetic order.

CS-wise, any arbitrary blank lines in the list of `use` statements have been removed.