[src] QA: use global constants with a fully qualified name
When PHP encounters an unqualified constant in a namespaced file, it will first try and find the constant within the namespace and only when it can not find it in the namespace, fall-back to the global constant.

Using fully qualified constant names when using a global constant ensures that PHP will bypass the search within the namespace and use the global constant directly.

Ref: https://www.php.net/manual/en/language.namespaces.fallback.php