WPSEO_Statistics: bug fix
This is primarily a bug in WP Core which states that `$found_posts` will be an integer, but then doesn't necessarily ensure that it is

All the same, this bug in WP Core does not need to have any effect in WPSEO and a simple cast to `int` will negate the effect of the bug.

Bug discovered while reviewing the unit tests.

Refs:
* https://developer.wordpress.org/reference/classes/wp_query/
* https://developer.wordpress.org/reference/classes/wp_query/set_found_posts/