CS: update Premium to use YoastCS 1.3.0
This is the sister-PR to the ruleset update pulled in the Premium repo.

This makes it so the `premium-check-cs` script will now use YoastCS 1.3.0.

It also ensures that files in the new `/tests/premium/` will be scanned and scanned against the correct PHP version (5.6+).

Lastly, it addresses the previously discussed issue that - as both repos currently still allow warnings -, errors can be hard to find in the report displayed by Travis. So for the Travis runs, CS `warning`s will now be hidden.