Add the translations of the javascript monorepo packages to the yoast-components textdomain
I presume that since we moved to the javascript monorepo there are some translations missing.
This adds the packages that actually have translations via `@wordpress/i18n` to the pipeline of `yoast-components`.
The textdomain is because that is how they are specified in the JS code. Perhaps later we should split this up properly.