Travis: only one command per condition
When there are several commands within the same condition, Travis will only look at the exit code of the last command to determine whether the build should pass/fail.

For commands not in the `script` section, this is normally not an issue as if something goes wrong there, the commands in `script` will likely fail too.

For the commands in the `script` section, however, we do want the build to fail, so we cannot have more than one command per condition.

The problem condition was the `$COVERAGE` check.
I have now:
* Moved the creating of the temporary directory to `before_script`.
* Split the remaining two unit test run commands and the command to merge the log files each into their own condition.