Travis: run PHP 7.4 against WP master
To get an honest impression about how PHP 7.4 ready the plugin is, the PHP 7.4 build needs to be run against the `trunk` version of WP (WP 5.3) as earlier WP versions weren't ready for PHP 7.4, so you would never get a passing build.

By running against `master`, any failures remaining should be PHP 7.4 issues in the plugin, not in WP itself.