Travis: run the "non-unit test" build against PHP 7.2
The "non-unit test" build was doing:
* Linting against PHP 7.3
* Running PHPCS
* Running a security check
* Running a JS check.

For the security check and the JS check, the PHP version doesn't matter.

For the PHPCS check, it also doesn't matter, other than that the PHPCS version used by WPSEO is a very old version which is not fully compatible with PHP 7.3, resulting in a lot of noise in the Travis output due to deprecations warnings and such coming from PHPCS.

With that in mind, I'm proposing changing the "non-unit test" build to run on PHP 7.2 which gets rid of the noise.

The only thing which was running on that build which really should be run against PHP 7.3, is the `lint` check, so moving that check to another PHP 7.3 build takes care of that.