Add method to retrieve posts with outdated words
Adds a method to retrieve posts with out of date prominent words. It
combines data from the indexables table and the posts table to
retrieve the posts. All posts that don't have indexables yet are
automatically included. This is because the subquery check to see that
there is an indexable with an up-to-date version. Then we select all
posts that don't have an indexable like this.