Tests: remove incorrect `@covers` tags
`@covers ClassName::<!public>` means that a test covers the methods in a class that are _not_ public.

Most of these tests are actually testing `public` methods, so the indicator is incorrect. Aside from that that, `<!public>` without class or `::` is incorrect as an annotation and PHPUnit has been complaining about this for a while (see detailed Travis output).