Decrease the time it takes to run yarn start
I did some simple tests (using time on my machine) and these were the results:
`yarn start` before: 15,10 seconds
`yarn start` after: 9,54 seconds
Which is a saving of 5 seconds, or 33% of the original startup time.

I did a quick smoketest through the plugin but I was not able to find anything out of the ordinary. No additional warnings or errors were thrown and everything I clicked workes as expected.