Remove the border-bottom and margin-top from Collapsible
The margin-top was a hack that prevented two borders from showing up when the collapsible were collapsed. With the changes to the tab container, this is no longer necesarry.

The outermost border is now added by the container, while the border top ensures that the collapsibles have a border when collapsed.