Fix error where $post->post_type only worked on posts
This problem was that we need to generic post type for the indexables table. So for all Pages, Custom Post types and Posts it should be "post".

What was happening is that for a post, it nicely returned post and the lookup succeeded. On custom post types, however, the function returned the type of the post (ie. "book"), which meant that the lookup did not succeed.

WPSEO_Utils::get_page_type() returns either "post" or "taxonomy" based on the parent type. This is exactly what we need!