Introduce thresholds check for CS
Bump PHPCS to ~2.0.0 to use the latest and greatest.
Introduces code that captures the PHPCS results output to have an opinion on the number of errors and warnings that are expected.
Updates the PHPCS configuration by removing exclusions that are marked to be irrelevant for the bumped version of YoastCS.
Fixes spaces vs tabs in PHPCS configuration.