Build/PHPCS: enable PHPCS 3.x configuration features
* Clean up the file paths PHPCS shows.
* Speed up by checking files in parallel.
* Enable caching between runs.

Regarding the caching feature:

Since PHPCS 3.0.0, PHPCS has a caching feature available. In this PR, this caching feature is now turned on for this repo.

This means in effect that the very first run of PHPCS with this feature will be marginally slower (on any individual machine). Most runs _after_ that though, will be significantly faster as PHPCS will only scan the files which have changed since the previous run.

The cache will automatically be invalidated and rebuild whenever relevant, like when any of the ruleset change, the version of one of the PHPCS dependencies has changed or when PHPCS itself has been changed.

To (temporarily) turn caching off, run PHPCS like so:
```bash
composer check-cs -- --no-cache
```

The cache file is set to be saved to a `.cache` directory to allow Travis to benefit from this feature as well.
This directory _has_ to exist for the cache file to be written to it, it is already created and committed to the repo with a `.gitkeep` file in a previous commit which enabled caching for Premium.