Build/PHPCS: *temporarily* ignore the filename and prefix rules for "old" code
Nearly all of the "old" code violates the filename rules and the prefix rules in YoastCS.

Fixing this should be a coordinated effort and preferably be done in one felt swoop instead of little by little as renaming files/classes/hooks etc is always a BC-break.

For that reason, these rules are - for now - only enabled for the "new" code, while we are making a planning for fixing this.