Composer: update PHP Scoper dependency
Update the `humbug/php-scoper` dependency from version `0.12.4` to `0.13.8`.

Based on the changelog, this should not make a significant difference for the task for which PHP Scoper is used by YoastSEO, aside from some bugfixes.

However, it does allow us to also update the underlying dependencies, getting rid of the last dependency which was blocking YoastSEO from being installed using Composer 2.x.

Ref:
* https://github.com/humbug/php-scoper/releases