Tests: use `expectOutputString()`
PHPUnit contains a native `expectOutputString()` method to set expectations for generated output.

If needs be, this can be combined with a call to `setOutputCallback()` method to adjust the actual generated output before the expectation is tested, for instance, to normalize line endings.

With that in mind, there is no need for a custom `expectOutput()` or `expectEmptyOutput()` methods in the `WPSEO_UnitTestCase` class.

So, this commit switches out the custom code in this test suite in favour of using the PHPUnit native `expectOutputString()` method.