Force_Rewrite_Title_Test: fix incorrect test
The `test_flush_cache_with_replacing_the_titles_after()` test is supposed to test the removal of the `<title>` tag _after_ the debug marker.

In reality it was testing removing the tag _before_ the debug marker, making it a duplicate of the `test_flush_cache_with_replacing_the_titles_before()` test.

This fixes the test input to test correctly what the test intended to test.