Schema_Blocks_Test: stabilize a test
The `Schema_Blocks::register_template()` expects either a relative or an absolute path and has a condition to handle relative paths.

This changes the `Schema_Blocks_Test::test_output()` to use a relative path instead of an absolute path, which makes sure the condition is tested, but also happens to stabilize the test on Windows as absolute paths on Windows, don't start with a `/` (but with a drive letter).