PHP 8.1 | HTML_Helper: add input validation
As of PHP 8.1, passing `null` to not explicitly nullable scalar parameters for PHP native functions is deprecated.

In the tests, a number of errors could be seen along the lines of:
```
preg_replace(): Passing null to parameter #3 ($subject) of type array|string is deprecated

path/to/src/helpers/schema/html-helper.php:30
path/to/src/generators/schema/website.php:31
path/to/src/generators/schema-generator.php:128
path/to/src/generators/schema-generator.php:76
path/to/tests/unit/generators/schema-generator-test.php:283
```

In this case, both methods in the `HTML_Helper` class expect a `string` as input, but it is not verified that they actually receive a `string` and the received parameter is blindly passed on to PHP native functions.
As running `strip_tags()` or the various regexes used with `preg_replace()` is useless, when the input is not a text string, as it wouldn't make a difference to the output, we can short circuit both functions by doing a type check at the start and bowing out early if the received parameter is not a string.

To maintain the "old" behaviour as much as possible:
* Integers and floats will be cast to string and returned (same as before as running these through the PHP native functions would also result in the parameter being cast to string).
* Stringable objects will be allowed through.
* In all other cases, an empty string is returned.
    This is a minor change in behaviour in case `true` would have been passed, which would previously return `1`, but IMO, that behaviour was incorrect anyway.
    This is also a change for when an array or (non-stringable) object would have been passed:
        - Previously an array or non-stringable object would lead to a PHP warning/error, now it will just return an empty string.

Includes:
* Adding tests safeguarding the support for stringable objects.
* Adding a test fixture for a stringable object.
    For now, this test fixture is added into this code base. At a later point in time, I'd like to propose to add typical "type test" fixture classes to WP Test Utils so these test utility classes are available to all projects.

Refs:
* https://wiki.php.net/rfc/deprecate_null_to_scalar_internal_arg
* https://www.php.net/manual/en/function.preg-replace.php